tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/106-module-import.html

<!DOCTYPE html>
<html><head>
        <title> scheduler: stylesheets blocking module scripts</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <!-- this stylesheet blocks scripts -->
        <link rel="stylesheet" href="css/import.css?pipe=trickle(d2)">
</head>
<body>
        <div id="log">FAILED (This TC requires JavaScript enabled)</div>
        <div id="test">Test</div>

        <script type="module">
          test(function() {
            assert_equals(getComputedStyle(document.getElementById("test")).position,
                          "fixed");
          });
        </script>
</body></html>
